+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed issue that variants missed by callVariant when multiple variants are causing the same sequence. #552
 
+- Fixed issue that `cpop_collapsed` attribute was not retained after merging so peptides that don't end with cleavage sites were yield. #554
+
 ---
 
 ## [0.9.2] - 2022-07-29